requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
